+++ pkg/webscan/options.go
@@ -9,6 +9,7 @@ type ScanOptions struct {
 	HTTPProtocol bool
 	HTTPHeader   bool
 	HTMLContent  bool
+	KnownFiles   bool
 	MailConfig   bool
 	Subdomain    bool
 }
@@ -21,6 +22,7 @@ func NewScanOptions(
 	httpProtocol bool,
 	httpHeader bool,
 	htmlContent bool,
+	knownFiles bool,
 	mailConfig bool,
 	subdomain bool,
 ) ScanOptions {
@@ -32,17 +34,18 @@ func NewScanOptions(
 		HTTPProtocol: httpProtocol,
 		HTTPHeader:   httpHeader,
 		HTMLContent:  htmlContent,
+		KnownFiles:   knownFiles,
 		MailConfig:   mailConfig,
 		Subdomain:    subdomain,
 	}
 }
 
 func AllScanOptions() ScanOptions {
-	return NewScanOptions(true, true, true, true, true, true, true, true, true)
+	return NewScanOptions(true, true, true, true, true, true, true, true, true, true)
 }
 
 func WebScanOptions() ScanOptions {
-	return NewScanOptions(true, true, true, true, true, true, true, false, false)
+	return NewScanOptions(true, true, true, true, true, true, true, true, false, false)
 }
 
 func (o ScanOptions) Apply(engine *Engine) {
@@ -66,6 +69,9 @@ func (o ScanOptions) Apply(engine *Engine) {
 	}
 	if o.HTMLContent {
 		engine.EnableHTMLContentScan()
+	}
+	if o.KnownFiles {
+		engine.EnableKnownFilesScan()
 	}
 	if o.MailConfig {
 		engine.EnableMailConfigScan()